[smart-release #195] refactor
